Added a --dry-run option to the licenser command
I have added a dry-run option to the Licenser command, this uses the existing
output messages to report affected files but will make no modifications to
any files if the option is passed to the command.

I have updated the output to include colour so that it's easier to read which
files are affected, and I have replaced occurrences of boolean with bool since
boolean is now deprecated. I have marked this feature (and another already
implemented feature) in the README as complete.